Move forwarder extension methods
