+++ src/ReverseProxy/Forwarder/IHttpForwarderExtensions.cs
@@ -19,7 +19,7 @@ namespace Yarp.ReverseProxy.Forwarder
         /// <param name="context">The HttpContext to forward.</param>
         /// <param name="destinationPrefix">The url prefix for where to forward the request to.</param>
         /// <param name="httpClient">The HTTP client used to forward the request.</param>
-        /// <returns>The result of forwarding the request and response.</returns>
+        /// <returns>The status of a forwarding operation.</returns>
         public static ValueTask<ForwarderError> SendAsync(this IHttpForwarder forwarder, HttpContext context, string destinationPrefix,
             HttpMessageInvoker httpClient)
         {
@@ -38,7 +38,7 @@ namespace Yarp.ReverseProxy.Forwarder
         /// <param name="destinationPrefix">The url prefix for where to forward the request to.</param>
         /// <param name="httpClient">The HTTP client used to forward the request.</param>
         /// <param name="requestConfig">Config for the outgoing request.</param>
-        /// <returns>The result of forwarding the request and response.</returns>
+        /// <returns>The status of a forwarding operation.</returns>
         public static ValueTask<ForwarderError> SendAsync(this IHttpForwarder forwarder, HttpContext context, string destinationPrefix,
             HttpMessageInvoker httpClient, ForwarderRequestConfig requestConfig)
         {
@@ -48,6 +48,35 @@ namespace Yarp.ReverseProxy.Forwarder
             }
 
             return forwarder.SendAsync(context, destinationPrefix, httpClient, requestConfig, HttpTransformer.Default);
+        }
+
+        /// <summary>
+        /// Forwards the incoming request to the destination server, and the response back to the client.
+        /// </summary>
+        /// <param name="context">The HttpContext to forward.</param>
+        /// <param name="destinationPrefix">The url prefix for where to forward the request to.</param>
+        /// <param name="httpClient">The HTTP client used to forward the request.</param>
+        /// <param name="requestTransform">Transform function to apply to the forwarded request.</param>
+        /// <returns>The status of a forwarding operation.</returns>
+        public static ValueTask<ForwarderError> SendAsync(this IHttpForwarder forwarder, HttpContext context, string destinationPrefix,
+            HttpMessageInvoker httpClient, Func<HttpContext, HttpRequestMessage, ValueTask> requestTransform)
+        {
+            return forwarder.SendAsync(context, destinationPrefix, httpClient, ForwarderRequestConfig.Empty, requestTransform);
+        }
+
+        /// <summary>
+        /// Forwards the incoming request to the destination server, and the response back to the client.
+        /// </summary>
+        /// <param name="context">The HttpContext to forward.</param>
+        /// <param name="destinationPrefix">The url prefix for where to forward the request to.</param>
+        /// <param name="httpClient">The HTTP client used to forward the request.</param>
+        /// <param name="requestConfig">Config for the outgoing request.</param>
+        /// <param name="requestTransform">Transform function to apply to the forwarded request.</param>
+        /// <returns>The status of a forwarding operation.</returns>
+        public static ValueTask<ForwarderError> SendAsync(this IHttpForwarder forwarder, HttpContext context, string destinationPrefix,
+            HttpMessageInvoker httpClient, ForwarderRequestConfig requestConfig, Func<HttpContext, HttpRequestMessage, ValueTask> requestTransform)
+        {
+            return forwarder.SendAsync(context, destinationPrefix, httpClient, requestConfig, new RequestTransformer(requestTransform));
         }
     }
 }
