+++ src/ReverseProxy/Forwarder/RequestTransformer.cs
@@ -1,13 +1,9 @@
-// Copyright (c) Microsoft Corporation.
-// Licensed under the MIT License.
-
 using System;
 using System.Net.Http;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Http;
-using Yarp.ReverseProxy.Forwarder;
 
-namespace Yarp.Sample
+namespace Yarp.ReverseProxy.Forwarder
 {
     internal class RequestTransformer : HttpTransformer
     {
